Split up test_cli by subcommand/option
